Added gulp for building the client side.
Added a node package called gulp for building the client side of the app. The build process is described in the gulpfile.js file. Currently it;

* Uglifies all javascript files in the public/src directory.
* Combinds the source javascript files with defined library files into a single file saved to `webapp/build/js/all.min.js`.
* Combinds all angular templates in the partials directory to `webapp/builds/js/template.js`.
* Minfies css files in public/css.
* Combinds the source css files and defined library files to a single file saved to `webapp/build/css/all.min.css`.
* Runs all images trough imgmin, compressing them for smaller file sizes.

To run a build of the client side one first has to install node (http://nodejs.org/).
Once installed, go to the webapp/public directory and run `npm install`.

The `npm install` command will;
1. Install all node dependencies (bower, gulp and some gulp plugins) locally in a folder called node_modules.
2. Run bower which will download libraries required by the client side (such as jQuery, angular, etc) to a folder called bower_components.
3. Run gulp, which will do what is mentioned above.

The npm install command is also integrated into maven, and will be run when maven does a packaging of the app. *This means node is required to build the project.*

This change also removes the bower_components folder from git. It is recommended to keep this folder in git, but some componenets ended up at too long file names, making windows go mad.

The maven configuration was also modified to only include the webapp/build directory in the .war, meaning the war will be smaller. Hopefully resulting in a faster deploy time.

TLDR;
Install node (http://nodejs.org/) and everything should just work. The first build will be very slow, as a lot of dependencies will need to be downloaded. After the first one, the added time should be reduced to a few seconds.