Added a dev-index for easier debugging.
Added a file "index-dev.jsp" served when going to /dev that uses non-compressed app resourses for easier debugging and "live" preview.

Going to /dev will work with auto-deploy on save. "/" is only updated on build. /dev still uses the lib.min.css and lib.min.js files, so a build will be required when adding a new 3rd party library.

Also note that /partials is mapped to public/partials. This mapping is only for the dev build, the production build overrides these via the partials.js file.