Added some tests.
These are based on the tests at https://github.com/spring-projects/spring-data-book/tree/master/jpa/src/test. Modified for our entitiy.

Also added a separate ApplicationConfig that is used by the tests (TestApplicationConfig.java) so that we can use an in-memory db for the tests and a "real" one for the app.