Changed some classes to be services.
Instead of defining @Bean factory methods, it turns out that one can also annotate classes with the @Service annotation. This should have the same effect as before (possible to inject via @Autowired) but makes more sense than having to declare a factory method for every class.

The classes that were made services are `TitleParser`, `OmdbHandler` and `FileThreadService`.

Some additional changes were made to OmdbHandler to make it use an injected RestTemplate class instead of creating one itself. This was to make mocking the RestTemplate instance possible.